Allow TCPStore to pick a port to bind to.
The motivation of this PR was to fix the problem where we would see
"Address already in use" issues for TCPStoreTest due to port conflicts. To
resolve this:

1. We can now pass in port 0 for TCPStore and retrieve the port it actually
bound to using a new getPort() API.
2. Added a `wait` flag to TCPStore constructor indicating whether or not it
should wait for workers (defaults to true).
3. Made `waitForWorkers` a public API to ensure that we can construct TCPStore
without waiting and wait for workers separately. This helps in TCPStoreTest to
ensure we can retrieve the port and pass it to the client stores.

Differential Revision: [D19240947](https://our.internmc.facebook.com/intern/diff/D19240947/)

ghstack-source-id: 96155025
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31674