[jit] More robust mangling
Summary:
Currently we keep a `mangleIndex_` that's intenral to compilation unit and
just increment the index when we found the original name is mangled, this doesn't
guarantee the new name is not defined.
This PR fixes the problem by querying whether the new name is defined or not.
fixes: https://github.com/pytorch/pytorch/issues/31268

Test Plan:
fixes the issue

Reviewers:
.

Subscribers:

Tasks:

Tags:

ghstack-source-id: 81c0d9c4d68b2da6a7e1d51dcee153435a203980
Pull Request resolved: https://github.com/pytorch/pytorch/pull/31978